change float to int

+++ timeoff_custom/report/hr_leave_report_balance.py
@@ -5,8 +5,8 @@ class HrLeaveReportBalance(models.Model):
     _auto = False
 
     name = fields.Char(string="Employee Name", readonly=True)
-    annual_leave = fields.Float(string="Annual Leave", readonly=True)
-    replacement_day_off = fields.Float(string="Replacement Day Off", readonly=True)
+    annual_leave = fields.Integer(string="Annual Leave", readonly=True)
+    replacement_day_off = fields.Integer(string="Replacement Day Off", readonly=True)
 
     def init(self):
         tools.drop_view_if_exists(self._cr, 'hr_leave_report_balance')
@@ -16,8 +16,8 @@ class HrLeaveReportBalance(models.Model):
                 select
                     row_number() OVER () as id,
                     balance.name,
-                    sum(case when balance.day_off_type ~* '(?<!\w)(?:annual)(?!\w)' then balance.balance_days end) as annual_leave,
-                    sum(case when balance.day_off_type ~* '(?<!\w)(?:replacement)(?!\w)' then balance.balance_days end) as replacement_day_off
+                    sum(case when balance.day_off_type ~* '(?<!\w)(?:annual)(?!\w)' then balance.balance_days end)::integer as annual_leave,
+                    sum(case when balance.day_off_type ~* '(?<!\w)(?:replacement)(?!\w)' then balance.balance_days end)::integer as replacement_day_off
                 from (
                     select
                         e.name,
